\name{clstutils-package}
\Rdversion{1.1}
\alias{clstutils-package}
\alias{clstutils}
\docType{package}
\title{
Sequence based classification and selection of reference sequences.
}

\description{Tools for performing taxonomic assignment based on phylogeny using pplacer and clst.}

\details{
\tabular{ll}{
Package: \tab clstutils\cr
Type: \tab Package\cr
Author: \tab Noah Hoffman <ngh2@uw.edu>\cr
License: \tab GPL3\cr
}

Index:
\preformatted{
}

% Further information is available in the following vignettes:
% \tabular{ll}{
% \code{clustrbDemo} \tab clstutils (source, pdf)\cr
% }

% TODO: write package overview.

}
\author{
Noah Hoffman

Maintainer: <ngh2@uw.edu>
}
% \references{
% }
\keyword{ package }
\keyword{ classif }
\seealso{
\code{\link{clst}}
}
\examples{
library(clstutils)
packageDescription("clstutils")
}
