\name{clstutils-package}
\Rdversion{1.1}
\alias{clstutils-package}
\alias{clstutils}
\docType{package}
\title{
Utilities for use with the clst2 package.
}

\description{Tools for performing taxonomic assignment based on phylogeny using pplacer and clst.}

\details{
\tabular{ll}{
Package: \tab clstutils\cr
Type: \tab Package\cr
Version: \tab 0.1.5256\cr
Date: \tab 2011-03-14\cr
Depends: \tab R (>= 2.10.0)\cr
Imports: \tab lattice, methods, RUnit\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
License: \tab UW Free Fork\cr
Built: \tab R 2.12.2; ; 2011-03-14; unix\cr
}

Index:
\preformatted{
}

% Further information is available in the following vignettes:
% \tabular{ll}{
% \code{clustrbDemo} \tab clstutils (source, pdf)\cr
% }

% TODO: write package overview.

}
\author{
Noah Hoffman

Maintainer: <ngh2@uw.edu>
}
% \references{
% }
\keyword{ package }
\keyword{ classif }
\seealso{
\code{\link{clst}}
}
\examples{
library(clstutils)
}
